+++ rl_envs/envs/labyrinth/display.py
@@ -4,7 +4,12 @@ from rl_envs.envs.labyrinth.constants import *
 
 
 class Display:
-    def __init__(self, rows, cols):
+    def __init__(self, rows: int, cols: int):
+        """
+        Args:
+            rows (int): The number of rows in the display.
+            cols (int): The number of columns in the display.
+        """
         self.screen = pygame.display.set_mode((CELL_SIZE * cols, CELL_SIZE * rows))
         pygame.display.set_caption("Labyrinth")
 
@@ -31,9 +36,10 @@ class Display:
         radius = CELL_SIZE // 2
 
         if cell_value == PLAYER:
-            pygame.draw.circle(self.screen, COLORS[PLAYER], (center_x, center_y), radius)
+            pygame.draw.circle(
+                self.screen, COLORS[PLAYER], (center_x, center_y), radius
+            )
         elif cell_value == TARGET:
             pygame.draw.circle(
                 self.screen, COLORS[TARGET], (center_x, center_y), radius
             )
-