+++ CHANGELOG.md
@@ -6,5 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
-- Built project structure
+ 
+## [0.1.0] - 01.09.2023
+### Added
+- Initial release of rl_envs - Labyrinth
+- Labyrinth - a 2D gridworld with randomly generated rectangular rooms, maze like paths, a starting position and a target position